// src/models/RoomModelType.ts
export interface RoomModelType {
    ma_phong:  number;
    ten_phong: string;
    khach:     number;
    phong_ngu: number;
    giuong:    number;
    phong_tam: number;
    mo_ta:     string;
    gia_tien:  number;
    may_giat:  boolean;
    ban_la:    boolean;
    tivi:      boolean;
    dieu_hoa:  boolean;
    bep:       boolean;
    do_xe:     boolean;
    ho_boi:    boolean;
    ban_ui:    boolean;
    ma_vi_tri: number;
    hinh_anh:  string;
}